Fix loading file with duplicates, improvements
Fixes a bug where loading a .zc file with duplicate entries would trip
a safety check. Cells are now properly deduplicated when loading files.
The last assignment to a cell is the one used.

Changes some indexes to usize instead of u32. More work needs to be done
here, as many choices of u32 for indexes weren't properly vetted. When
using a non-usize index, overflow cases need to be explicitly handled.
This wasn't done for most cases in the codebase, so either overflow
checks need to be done or indexes should be changed to usize. Certain
indexes should have the ability to be larger than u32, such as cell
handles or ast node indexes.

Now uses bulk cell delete undos when loading a sheet, instead of
creating a single delete cell undo per cell loaded. This reduces memory
usage and makes undoing and redoing a sheet load much faster.

Now counts columns properly when loading sheets. Previous implementation
would allocate `n` columns, where `n` is the number of cells being
loaded.

Removed the unused string index field from cells. It wasn't doing
anything since string indexes had been moved into the individual AST
nodes a while ago.

Reduces memory usage of the cell string buffer. Remnants of the old
implementation using string indexes stored in cells caused each cell
assignment to append a zero byte to the buffer for null termination.
This has not been needed for a while and is now removed. The string
buffer is now only appended to when we actually encounter string
literals.

+++ src/Parser.zig
@@ -14,7 +14,7 @@ token_starts: []const u32,
 tok_i: u32,
 
 /// Total byte length of all parsed string literals.
-strings_len: u32,
+strings_len: usize,
 
 src: []const u8,
